resolve PR
